requirements.txt
Django~=5.2.1
django-ckeditor-5~=0.2.18
django-stubs[compatible-mypy]~=5.2.1
django-tailwind[reload]~=4.0.1
djangorestframework~=3.16.0
djangorestframework-stubs[compatible-mypy]~=3.16.0
drf-nested-routers~=0.94.2
drf-spectacular~=0.28.0
mypy~=1.15.0